refactor: relocate shared decode kernels + sampling to DecodeKernels.swift (P5)
Pure numeric kernels (multiplyRowMajorMatrix, projectRowMajorMatrix,
rmsNorm, applyHalfSplitRoPEPerHead, silu, addBiasInPlace) and the
sampling family (partitionedArgmax, evaluateGreedyClassifier,
greedyArgmax) move out of the engine monolith into
Sources/RealModelInference/DecodeKernels.swift behind the frozen engine
extension. No behavior change.

Verified: golden replay bit-identical on all four trunks; full suite green.
